requirements.txt
django
djangorestframework
django-filter
markdown
qrcode
